% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllCombos.R
\name{AllCombos}
\alias{AllCombos}
\title{AllCombos}
\usage{
AllCombos(alltokens)
}
\arguments{
\item{alltokens}{A data.frame containing model features tokens}
}
\description{
Create all possible combinations of tokens sets. alltokens argument must be 
 a dataframe with $tokengroup, $tokenset, $token
}
