\name{Apkg-package}
\alias{Apkg-package}
\alias{Apkg}
\docType{package}
\title{
\packageTitle{Apkg}
}
\description{
\packageDescription{Apkg}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{Apkg}

Maintainer: \packageMaintainer{Apkg}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = Apkg)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{Apkg}}
%\section{Documentation Index}{\packageIndices{Apkg}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
