% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateModel.R
\name{CheckThenCreate}
\alias{CheckThenCreate}
\title{CheckThenCreate}
\usage{
CheckThenCreate(path, control = NULL, alltokens = NULL, allmods = NULL)
}
\arguments{
\item{path}{A character string containing the path to control stream.}

\item{control}{Optional. Control stream template if model control stream does not exist yet.}

\item{alltokens}{Optional. data frame containing tokens.}

\item{allmods}{Optional. Data frame containing all model combinations.}
}
\description{
This function checks if a model control stream exists, and creates it based on control stream template and tokens if it doesn't
}
