% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateModel.R
\name{CopyModel}
\alias{CopyModel}
\title{Copy Control Stream}
\usage{
CopyModel(copy, copyTo, control = NULL, alltokens = NULL, allmods = NULL)
}
\arguments{
\item{copy}{A character string containing the path to control stream.}

\item{copyTo}{A character string containing the directory to copy to.}

\item{control}{Optional. Control stream template if model control stream does not exist yet.}

\item{alltokens}{Optional. data frame containing tokens.}

\item{allmods}{Optional. Data frame containing all model combinations.}
}
\description{
This function copies a control stream to a new directory
}
