% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateModel.R
\name{CreateModel}
\alias{CreateModel}
\title{CreateModel}
\usage{
CreateModel(controlstream, alltokens, phenotype, tokengroups)
}
\arguments{
\item{controlstream}{A character string of control stream template.}

\item{alltokens}{Data frame containing tokens.}

\item{phenotype}{Data frame containing specific model features (tokensets).}

\item{tokengroups}{Data frame containing general model features (tokengroups).}
}
\description{
Creates syntactically correct control stream based on template and tokens
CreateModel takes a control stream, a token data frame (used as a look up)
to translate phenotypes to model syntax, a model phenotype, and tokengroups
it returns a new control stream
}
