% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneticOperators.R
\name{crossover}
\alias{crossover}
\title{crossover}
\usage{
crossover(selectedmods, ncrossoverpoints)
}
\arguments{
\item{selectedmods}{Character vector of paths to selected models}
}
\value{
Returns data.frame of model phenotypes
}
\description{
Performs n-point crossover on selected individuals
}
