% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\name{fitness}
\alias{fitness}
\title{fitness}
\usage{
fitness(results = data.frame(), penalties = list(THETA = 5, ETA = 5, EPS =
  5, cov.success = -10, min.success = -10, cov.warnings = 0, boundary = 0,
  rounding = 0, zero.grad = 0, final.zero.grad = 0, hessian.reset = 0,
  s.singular = 0, sig.digs = 0, condition.number = 0))
}
\arguments{
\item{results}{data.frame containing the results of a model run}

\item{penalties}{List containing penalty coefficients}
}
\description{
Determine the fitness of a model run
}
