% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello.gas_planet}
\alias{hello.gas_planet}
\title{Say hello to a gas planet}
\usage{
\method{hello}{gas_planet}(planet, ...)
}
\arguments{
\item{planet}{a planet object}

\item{...}{More arguments}
}
\description{
Say hello to a gas planet
}
\examples{
planet <- set_gas_planet("Jupiter")
hello(planet)

}
