% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitiaiteGA.R
\name{InitiateGA}
\alias{InitiateGA}
\title{InitiateGA}
\usage{
InitiateGA(nmods, nindiv, control = NULL, alltokens = NULL,
  allmods = NULL, seed = 1)
}
\arguments{
\item{nmods}{Numeric. The number of all possible models}

\item{nindiv}{Numeric. The number of individuals per generation}

\item{control}{Optional. Control stream template if model control stream does not exist yet.}

\item{alltokens}{Optional. data frame containing tokens.}

\item{allmods}{Optional. Data frame containing all model combinations.}

\item{seed}{Numeric. Seed number for reproducibility}
}
\description{
Initiate the genetic algorithm on defined model features
}
