% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneticOperators.R
\name{mutation}
\alias{mutation}
\title{Mutation}
\usage{
mutation(crossoveredmods, alltokens, options = list(pmutation = 0.07))
}
\arguments{
\item{crossoveredmods}{Data frame of models from crossover step}

\item{alltokens}{Data frame containing tokens.}

\item{options}{List containing mutation options (pmutation)}
}
\value{
Returns data.frame of model phenotypes
}
\description{
Performs mutation
}
