% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitiaiteGA.R
\name{NextGA}
\alias{NextGA}
\title{NextGA}
\usage{
NextGA(alltokens, control = NULL, allmods = NULL)
}
\arguments{
\item{alltokens}{Data frame containing tokens.}

\item{control}{Control stream template if model control stream does not exist yet.}

\item{allmods}{Optional. Data frame containing all model combinations.}

\item{mods}{Character vector of paths to candidate models}
}
\description{
Determine the next generation of individuals
}
