% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartments.R
\name{p1cmt}
\alias{p1cmt}
\title{p1cmt}
\usage{
p1cmt(compartment, cmtn, ncmts)
}
\arguments{
\item{compartment}{Character string containing name of center compartment}

\item{cmtn}{Numeric. Compartment number of center compartment}

\item{ncmts}{Numeric. Number of compartments already present}
}
\description{
Create 1 compartment diagram
}
