% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartments.R
\name{p3cmt}
\alias{p3cmt}
\title{p3cmt}
\usage{
p3cmt(compartment, cmtn, ncmts)
}
\arguments{
\item{compartment}{Character string containing name of center compartment}

\item{cmtn}{Numeric. Compartment number of center compartment}

\item{ncmts}{Numeric. Number of compartments already present}
}
\description{
Create 3 compartment diagram
}
