% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgA.R
\docType{package}
\name{pkgA}
\alias{pkgA}
\title{pkgA for demonstration purpose of linking native routines}
\description{
pkgA for demonstration purpose of linking native routines
}
\author{
Jitao David Zhang <jitao_david.zhang@roche.com>
}
