% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgA.R
\name{pkgAfunc}
\alias{pkgAfunc}
\title{A dummy function exported for demonstration purposes}
\usage{
pkgAfunc(username)
}
\arguments{
\item{username}{Username in characters}
}
\value{
A number generated in C
}
\description{
A dummy function exported for demonstration purposes
}
