% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressionModels.R
\name{powmod}
\alias{powmod}
\title{powmod}
\usage{
powmod(x, y, xlab = "X", ylab = "Y")
}
\arguments{
\item{x}{Numeric vector}

\item{y}{Numeric vector}

\item{xlab}{Character string. X variable label.}

\item{ylab}{Character string. Y variable label.}
}
\description{
Linear Regression between log(x) and log(y). Returns data.frame to be displayed in UI
containing relationship (i.e power), model (ie WTonCL), estimate and pval.
}
