% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RetrieveResults.R
\name{RetrieveResults}
\alias{RetrieveResults}
\title{RetrieveResults}
\usage{
RetrieveResults(directory)
}
\arguments{
\item{directory}{Character string containing project directory}
}
\description{
Retrieve results if results directory exists. returns dataframe with model number,
OFV, run success status, and covariance success status
This version of retrieveresults is used for the "Allmods" section
}
