% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RetrieveResults.R
\name{RetrieveResultsEach}
\alias{RetrieveResultsEach}
\title{RetrieveResultsEach}
\usage{
RetrieveResultsEach(directory)
}
\arguments{
\item{directory}{Character string containing project directory}
}
\description{
Retrieve results if results directory exists. returns dataframe with model number,
OFV, run success status, and covariance success status
This version of retrieve results is used in the GA portion of the interface,
as well as user created subdirectories outside of "All".
}
