% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel.R
\name{RunModel}
\alias{RunModel}
\title{RunModel}
\usage{
RunModel(directory, model = NULL, wait = T, exit = T)
}
\arguments{
\item{directory}{A character string containing the path to folder in which mod.ctl is located}

\item{model}{Optional. A character string providing title of running process}
}
\description{
This function executes mod.ctl in the path specified via PsN.
}
