% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneticOperators.R
\name{selection}
\alias{selection}
\title{Selection}
\usage{
selection(candidates)
}
\arguments{
\item{candidates}{Character vector of paths to candidate models}
}
\value{
Returns character vector of paths to selected models
}
\description{
Performs tournament style selection
}
